fill_packs_from_git: fix leaking of temporary files
In this function, all opened temporary files were left
undeleted and became garbage in /tmp.
CLEANUP is not valid flag for tempfile().
For tempfile, the flag would be UNLINK, but it can't be used
when OPEN is set to zero (which is the case).
Perl will not automatically delete such files,
so we have to close and delete them.

Change-Id: Id05ffa5c41f9d7522faa370fb8816ebb64740be6